splits "test_source.py" into multiple files
